finished referrer in url, moving to client-side query
