+++ test/test_helper.rb
@@ -86,7 +86,7 @@ module TestHelpers
                                referrer: Referrer.find_or_create_by(address: "http://amazon.com"),
                                request_type: RequestType.find_or_create_by(verb: "GET"),
                                event: Event.find_or_create_by(name: "facebook"),
-                               u_agent: UAgent.find_or_create_by(browser: "Chrome", platform: "Webkit"),
+                               u_agent: UAgent.find_or_create_by(browser: "Opera", platform: "Webkit"),
                                resolution: Resolution.find_or_create_by(width: "1920", height: "1280"),
                                ip: Ip.find_or_create_by(address: "63.29.38.211"),
                                requested_at: "2013-02-16 21:37:00 -0700",
@@ -96,7 +96,7 @@ module TestHelpers
                                referrer: Referrer.find_or_create_by(address: "http://newegg.com"),
                                request_type: RequestType.find_or_create_by(verb: "POST"),
                                event: Event.find_or_create_by(name: "twitter"),
-                               u_agent: UAgent.find_or_create_by(browser: "Safari", platform: "Macintosh"),
+                               u_agent: UAgent.find_or_create_by(browser: "Mozilla", platform: "Windows"),
                                resolution: Resolution.find_or_create_by(width: "1920", height: "1280"),
                                ip: Ip.find_or_create_by(address: "63.29.38.200"),
                                requested_at: "2013-02-16 21:38:00 -0700",
